+++ Assets/_Scripts/GameManager.cs
@@ -4,6 +4,8 @@ namespace _Scripts
 {
     public class GameManager : MonoBehaviour
     {
+
+        // [SerializeField] private GameObject player;
         private static GameManager gameManager { get; set; }
 
         // Start is called before the first frame update
@@ -17,7 +19,7 @@ namespace _Scripts
             {
                 gameManager = this;
             }
-
+            // Instantiate(player, new Vector3(115, 1, -65), Quaternion.identity);
         }
 
         private void Start()